Conflicts:
	src/main/java/com/foundationdb/qp/persistitadapter/PersistitAdapter.java
	src/main/java/com/foundationdb/qp/persistitadapter/indexrow/PersistitIndexRow.java
	src/test/java/com/foundationdb/qp/operator/OperatorTestHelper.java

+++ src/test/java/com/foundationdb/server/api/dml/scan/NiceRowTest.java
@@ -46,7 +46,7 @@ public final class NiceRowTest {
 
         // Why -1: because an __akiban_pk column gets added
         assertEquals("fields count", 2, newRow.getFields().size() - 1);
-        assertEquals("field[0]", 5L, newRow.get(0));
+        assertEquals("field[0]", 5, newRow.get(0));
         assertEquals("field[1]", "Bob", newRow.get(1));
 
         compareRowDatas(rowData, newRow.toRowData());
@@ -71,10 +71,10 @@ public final class NiceRowTest {
 
         // Why -1: because an __akiban_pk column gets added
         assertEquals("fields count", NUM, newRow.getFields().size() - 1);
-        assertEquals("field[0]", 15L, newRow.get(0));
+        assertEquals("field[0]", 15, newRow.get(0));
         assertEquals("field[1]", "Robert", newRow.get(1));
         for (int i=2; i < NUM; ++i) {
-            long expected = i + 1000;
+            int expected = i + 1000;
             assertEquals("field[1]", expected, newRow.get(i));
         }
 
@@ -107,10 +107,10 @@ public final class NiceRowTest {
 
         // Why -1: because an __akiban_pk column gets added
         assertEquals("fields count", NUM, newRow.getFields().size() - 1);
-        assertEquals("field[0]", 15L, newRow.get(0));
+        assertEquals("field[0]", 15, newRow.get(0));
         assertEquals("field[1]", "Robert", newRow.get(1));
         for (int i=2; i < NUM; ++i) {
-            Long expected = (i % 3) == 0 ? null : i + 1000L;
+            Integer expected = (i % 3) == 0 ? null : i + 1000;
             assertEquals("field[1]", expected, newRow.get(i));
         }
 
@@ -147,12 +147,12 @@ public final class NiceRowTest {
         String ddl = "create table test.t(id int not null primary key, v varchar(255) character set utf8)";
         RowDef rowDef = SCHEMA_FACTORY.aisWithRowDefs(ddl).getTable("test", "t").rowDef();
 
-        Object[] objects = { 1L, str };
+        Object[] objects = { 1, str };
         RowData rowData = create(rowDef, objects);
         NewRow newRow = NiceRow.fromRowData(rowData, rowDef);
 
         assertEquals("fields count", 2, newRow.getFields().size());
-        assertEquals("field[0]", 1L, newRow.get(0));
+        assertEquals("field[0]", 1, newRow.get(0));
         assertEquals("field[1]", str, newRow.get(1));
         assertEquals("filed[1] charset", "utf8", rowDef.getFieldDef(1).column().getCharsetAndCollation().charset());
 
@@ -172,12 +172,12 @@ public final class NiceRowTest {
         String ddl = "create table test.t(id int not null primary key, v varchar(32) character set utf8)";
         RowDef rowDef = SCHEMA_FACTORY.aisWithRowDefs(ddl).getTable("test", "t").rowDef();
 
-        Object[] objects = { 1L, TEST_STR };
+        Object[] objects = { 1, TEST_STR };
         RowData rowData = create(rowDef, objects);
         NewRow newRow = NiceRow.fromRowData(rowData, rowDef);
 
         assertEquals("fields count", 2, newRow.getFields().size());
-        assertEquals("field[0]", 1L, newRow.get(0));
+        assertEquals("field[0]", 1, newRow.get(0));
         assertEquals("field[1]", TEST_STR, newRow.get(1));
         assertEquals("filed[1] charset", "utf8", rowDef.getFieldDef(1).column().getCharsetAndCollation().charset());
